# src/swagger/vj-ota-api.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1",
    "title": "VietJet API GATEWAY",
    "description": "VietJet API GATEWAY",
    "termsOfService": "None",
    "contact": {
      "name": "VietJet API GATEWAY"
    },
    "license": {
      "name": "All right reserved"
    }
  },
  "servers": [
    {
      "url": "https://34.144.235.95.nip.io/flight"
    }
  ],
  "paths": {
...
    "/countries": {
      "get": {
        "tags": ["Countries"],
        "summary": "Retrieve a list of all countries.",
        "operationId": "ApiCountriesGet",
        "parameters": [
          {
            "name": "includeProvinces",
            "in": "query",
            "description": "Optional inclusion of provinces.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "uniqueItems": false,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Invalid Credentials": {
                      "type": "string",
                      "example": "Occurs when the api username and/or password are not valid for the endpoint you are trying to reach"
                    },
                    "Permission Denied": {
                      "type": "string",
                      "example": "Occurs when you do not have permission to perform the target action"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/countries/{code}/provinces": {
      "get": {
        "tags": ["Countries"],
        "summary": "Retrieve a list of provinces in a specific country.",
        "operationId": "ApiCountriesByCodeProvincesGet",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "uniqueItems": false,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Invalid Credentials": {
                      "type": "string",
                      "example": "Occurs when the api username and/or password are not valid for the endpoint you are trying to reach"
                    },
                    "Permission Denied": {
                      "type": "string",
                      "example": "Occurs when you do not have permission to perform the target action"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/userSessions": {
      "post": {
        "tags": ["UserSessions"],
        "summary": "Get access token",
        "operationId": "ApiUserSessionsPost",
        "security": [
          {
            "basicAuth": [],
            "api-key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSessionModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Invalid Credentials": {
                      "type": "string",
                      "example": "Occurs when the api username and/or password are not valid for the endpoint you are trying to reach"
                    },
                    "Permission Denied": {
                      "type": "string",
                      "example": "Occurs when you do not have permission to perform the target action"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["UserSessions"],
        "summary": "Refresh access token",
        "operationId": "ApiUserSessionsPut",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                  },
                  "refreshToken": {
                    "type": "string",
                    "example": "UxDSmxibU1pT2lKQk1qVTJRMEpETFVoVE5UR"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSessionModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Invalid Credentials": {
                      "type": "string",
                      "example": "Occurs when the api username and/or password are not valid for the endpoint you are trying to reach"
                    },
                    "Permission Denied": {
                      "type": "string",
                      "example": "Occurs when you do not have permission to perform the target action"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "api-key": {
        "type": "apiKey",
        "name": "apikey",
        "in": "header"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "This is method for login user"
      }
    },
    "schemas": {
...
      "CountryModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "provinces": {
            "$ref": "#/components/schemas/ProvinceModel"
          }
        }
      },
      "ProvinceModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UserSessionModel": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "api-key": []
    }
  ]
}